th-27: * fixed all the issues from the comments
