+++ frontend/.env.example
@@ -5,3 +5,6 @@ VITE_APP_API_ORIGIN_URL=/api/v1
 
 # DEV
 VITE_APP_PROXY_SERVER_URL=http://localhost:3001
+
+#APi
+VITE_APP_GOOGLE_MAPS_API_KEY=your_google_maps_api_key
