+++ frontend/src/libs/components/app/app.tsx
@@ -16,7 +16,10 @@ const App: React.FC = () => {
     users: users.users,
     dataStatus: users.dataStatus,
   }));
-
+  const mapProperties = {
+    center: { lat: 50.4501, lng: 30.5234 },
+    zoom: 10,
+  };
   const isRoot = pathname === AppRoute.ROOT;
 
   useEffect(() => {
@@ -50,7 +53,7 @@ const App: React.FC = () => {
           <h2>Users:</h2>
           <h3>Status: {dataStatus}</h3>
           <div style={{ width: '500px', height: '500px' }}>
-            <Map />
+            <Map center={mapProperties.center} zoom={mapProperties.zoom} />
           </div>
           <ul>
             {users.map((it) => (
