+++ frontend/src/libs/components/map/map-inner-component/map-inner-component.tsx
@@ -1,19 +1,16 @@
 import { useEffect, useRef } from '~/libs/hooks/hooks.js';
-import { MapService } from '~/libs/packages/map/map.package';
+import { MapService } from '~/libs/packages/map/map.js';
+import  { type MapType } from '~/libs/types/types.js';
 
 import styles from './map.module.scss';
 
-type Properties = {
-  center: google.maps.LatLngLiteral;
-  zoom: number;
-  origin?: google.maps.LatLngLiteral;
-  destination?: google.maps.LatLngLiteral;
+type Properties = MapType & {
   className?: string;
 };
 
 const MapInnerComponent: React.FC<Properties> = ({
   center,
-  zoom,
+  zoom, 
   className,
 }: Properties) => {
   const mapReference = useRef<HTMLDivElement>(null);
