+++ frontend/src/libs/packages/map/map.ts
@@ -1,0 +1,2 @@
+
+export { MapService } from './map.package.js' ;
