+++ frontend/src/libs/types/map.type.ts
@@ -1,0 +1,8 @@
+type MapType = {
+    center: google.maps.LatLngLiteral;
+    zoom: number;
+    origin?: google.maps.LatLngLiteral;
+    destination?: google.maps.LatLngLiteral;
+  };
+
+  export { type MapType };
