+++ frontend/src/libs/types/types.ts
@@ -1,5 +1,6 @@
 export { type AsyncThunkConfig } from './async-thunk-config.type.js';
 export { type FormField } from './form.type.js';
+export { type MapType } from './map.type.js';
 export { type DeepPartial, type FieldValues } from 'react-hook-form';
 export {
   type ServerErrorDetail,
